chore: examples/react/ui -> examples/react/playground
License: MIT
