+++ examples/react/playground/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>W3UI React Components</title>
+    <title>W3UI React Playground</title>
     </head>
   <body style="background-color:#1d2027;">
     <div id="root"></div>
