+++ examples/react/playground/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@w3ui/example-react-ui",
+  "name": "@w3ui/example-react-playground",
   "private": true,
   "version": "0.0.0",
   "type": "module",
