+++ examples/react/playground/public/favicon.ico
