+++ examples/react/playground/src/components/Authenticator.jsx
