+++ packages/react/src/index.ts
@@ -1,5 +1,5 @@
-export * from './SimpleUploader'
-export * from './SimpleUploadsList'
-export * from './SimpleAuthenticator'
+export * from './Uploader'
+export * from './UploadsList'
+export * from './Authenticator'
 export * from './W3Upload'
 export * from './providers/W3API'
